pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "depswiz"
version = "0.5.0"
description = "Multi-language dependency wizard - check, audit, and update dependencies across Python, Rust, Dart, and JavaScript ecosystems"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [{ name = "depswiz contributors" }]
keywords = [
    "dependencies",
    "security",
    "vulnerability",
    "sbom",
    "audit",
    "python",
    "rust",
    "dart",
    "javascript",
    "typescript",
    "monorepo",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]

dependencies = [
    # CLI Framework
    "typer>=0.15.0",
    "rich>=13.9.0",

    # HTTP Client (async)
    "httpx>=0.27.0",

    # Parsing
    "pyyaml>=6.0.0",

    # Version handling
    "packaging>=24.0",
    "semver>=3.0.0",

    # SBOM generation
    "cyclonedx-python-lib>=11.0.0",

    # Async support
    "anyio>=4.0.0",

    # Caching
    "diskcache>=5.6.0",

    # TUI Framework (for guide command)
    "textual>=0.89.0",

    # Interactive prompts (for wizard mode)
    "inquirerpy>=0.3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]
docs = ["mkdocs>=1.6.0", "mkdocs-material>=9.5.0"]

[project.scripts]
depswiz = "depswiz.cli:app"

[project.entry-points."depswiz.languages"]
python = "depswiz.plugins.python:PythonPlugin"
rust = "depswiz.plugins.rust:RustPlugin"
dart = "depswiz.plugins.dart:DartPlugin"
javascript = "depswiz.plugins.javascript:JavaScriptPlugin"
docker = "depswiz.plugins.docker:DockerPlugin"
golang = "depswiz.plugins.golang:GoPlugin"

[project.urls]
Homepage = "https://github.com/moinsen-dev/depswiz"
Documentation = "https://moinsen-dev.github.io/depswiz"
Repository = "https://github.com/moinsen-dev/depswiz"
Issues = "https://github.com/moinsen-dev/depswiz/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/depswiz"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (required for typer.Option/Argument)
    "ARG001", # unused function argument (often intentional in interfaces)
    "ARG002", # unused method argument (often intentional in interfaces)
    "SIM102", # nested if statements (sometimes clearer)
    "SIM108", # ternary operator (sometimes less readable)
    "SIM105", # contextlib.suppress (try-except-pass is clearer)
    "SIM116", # dictionary instead of if-elif (less readable for this case)
    "SIM110", # any() instead of for loop (less readable)
    "PTH123", # Path.open() vs open() (both are fine)
    "ERA001", # commented-out code (intentional TODOs)
    "B028",   # stacklevel in warnings (not critical)
]

[tool.ruff.lint.isort]
known-first-party = ["depswiz"]

[tool.ruff.lint.per-file-ignores]
# typer.Exit() in exception handlers is intentional control flow
"src/depswiz/cli/commands/*.py" = ["B904"]
# Guide module uses runtime types needed by Textual and dataclasses
"src/depswiz/guide/*.py" = ["TC002", "TC003", "RUF012"]
"src/depswiz/guide/screens/*.py" = ["TC002", "TC003", "RUF012"]
# Deprecations module uses Path in dataclass fields
"src/depswiz/deprecations/*.py" = ["TC003"]

[tool.mypy]
python_version = "3.13"
# Relaxed settings for practical development
strict = false
warn_return_any = false
warn_unused_ignores = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
# Still check for important issues
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=depswiz --cov-report=term-missing"

[tool.coverage.run]
source = ["src/depswiz"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
